reorganize the cli export into multiple files
