--- modality-probe-cli/src/graph/mod.rs
@@ -1,3 +1,0 @@
-mod meta;
-
-pub(crate) use crate::graph::meta::*;
