+++ modality-probe-cli/src/lib.rs
@@ -2,7 +2,6 @@ pub mod component;
 pub mod error;
 pub mod events;
 pub mod export;
-pub mod graph;
 pub mod header_gen;
 pub mod lang;
 pub mod manifest_gen;
